+++ lib/helpers/index.ts
@@ -1,3 +1,3 @@
-export * from './listBuilder'
-export * from './truncate'
-export * from './StringCapitalizer'
+export * from './listBuilder';
+export * from './truncate';
+export * from './stringCapitalizer';
