+++ lib/hooks/useSignInWithGoogle.ts
@@ -1,0 +1,12 @@
+
+import { auth,  googleAuthProvider } from '@/lib/firebase';
+
+
+
+export const useSignInWithGoogle = async () => {
+  try {
+    await auth.signInWithPopup(googleAuthProvider);
+  } catch (err) {
+    console.log(err);
+  }
+};
