+++ styles/GlobalStyle.ts
@@ -1,6 +1,7 @@
 import { createGlobalStyle } from 'styled-components';
 
 const GlobalStyle = createGlobalStyle`
+
 html {
   scroll-behavior: smooth;
   background-color: rgb(249, 250, 251)
@@ -18,7 +19,6 @@ html {
   }
   &::-webkit-scrollbar-thumb {
     background: rgba(52, 211, 153);
-    /* border: 0.25em solid hsl(120 75% 50% / 1); */
     border-radius: 100vw;
   }
 @supports (scrollbar-color: rgba(52, 211, 153) rgba(156, 163, 175)) {
