--- testData/exampleSearchResponse.js
@@ -1,40 +1,0 @@
-export const dummyData = {
-  results: [
-    {
-      id: 646512,
-      title: 'Salmon Caesar Salad',
-      image: 'https://spoonacular.com/recipeImages/646512-312x231.jpg',
-      imageType: 'jpg',
-    },
-    {
-      id: 638166,
-      title: 'Chicken Liver Salad',
-      image: 'https://spoonacular.com/recipeImages/638166-312x231.jpg',
-      imageType: 'jpg',
-    },
-    {
-      id: 1095697,
-      title: '15-Minute Tamari Marinated Steak Salad ',
-      image: 'https://spoonacular.com/recipeImages/1095697-312x231.jpg',
-      imageType: 'jpg',
-    },
-    {
-      id: 638550,
-      title: 'Chili Chicken Salad',
-      image: 'https://spoonacular.com/recipeImages/638550-312x231.jpg',
-      imageType: 'jpg',
-    },
-    {
-      id: 1095693,
-      title: 'Raspberry Arugula Side Salad ',
-      image: 'https://spoonacular.com/recipeImages/1095693-312x231.jpg',
-      imageType: 'jpg',
-    },
-    {
-      id: 650378,
-      title: 'Curry Chicken Salad',
-      image: 'https://spoonacular.com/recipeImages/650378-312x231.jpg',
-      imageType: 'jpg',
-    },
-  ],
-};
