--- testData/placeholderFavorites.js
@@ -1,62 +1,0 @@
-export const placeholderFavorites = [
-  {
-    recipeId: 1096010,
-    title: 'Egg Salad Wrap',
-  },
-  {
-    recipeId: 642240,
-    title: 'Egg Salad Sandwiches With Tarragon',
-  },
-  {
-    recipeId: 637898,
-    title: 'Chicken and Dumpling Soup',
-  },
-  {
-    recipeId: 643933,
-    title: 'Fruid Salad Dressing',
-  },
-  {
-    recipeId: 157106,
-    title: 'Wedge Salad',
-  },
-  {
-    recipeId: 656329,
-    title: 'Pizza bites with pumpkin',
-  },
-  {
-    recipeId: 680975,
-    title: 'BLT Pizza',
-  },
-  {
-    recipeId: 663136,
-    title: 'Thai Pizza',
-  },
-  {
-    recipeId: 716300,
-    title: 'Plantain Pizza',
-  },
-  {
-    recipeId: 665769,
-    title: 'Zucchini Pizza Boats',
-  },
-  {
-    recipeId: 655698,
-    title: 'Pepperoni Pizza Muffins',
-  },
-  {
-    recipeId: 622598,
-    title: 'Pittata - Pizza Frittata',
-  },
-  {
-    recipeId: 641893,
-    title: 'Easy Cheesy Pizza Casserole',
-  },
-  {
-    recipeId: 655847,
-    title: 'Pesto Veggie Pizza',
-  },
-  {
-    recipeId: 654523,
-    title: 'Paneer & Fig Pizza',
-  },
-];
