+++ tsconfig.json
@@ -6,7 +6,7 @@
       "@/styles/*": ["styles/*"],
       "@/contexts/*": ["contexts/*"],
       "@/lib/*": ["lib/*"],
-      "@/types/*": ["types/*"],
+      "@/types/*": ["types/index.ts"],
     },
     "target": "es5",
     "lib": [
@@ -30,7 +30,7 @@
     "next-env.d.ts",
     "**/*.ts",
     "**/*.tsx"
-  ],
+, "pages/_document.js"  ],
   "exclude": [
     "node_modules"
   ]
